+++ docs/release-notes.md
@@ -1,5 +1,7 @@
 ## Latest changes
 
+* Fix bug with ordering of routes in a CBV router
+
 ## 0.1.1
 
 * Add source docstrings for most functions.